unify assertEqualPlans for join reorder related suites
